% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapevine-package.R
\docType{package}
\name{grapevine-package}
\alias{grapevine}
\alias{grapevine-package}
\title{grapevine: The Grapevine}
\description{
A simple set of useful, unique, and non-standard binary operators.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stufield/grapevine}
}

}
\author{
\strong{Maintainer}: Stu Field \email{stu.g.field@gmail.com} (\href{https://orcid.org/0000-0002-1024-5859}{ORCID}) [copyright holder]

}
\keyword{internal}
