% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapes.R
\name{grapes}
\alias{grapes}
\alias{\%!in\%}
\alias{\%||\%}
\alias{\%==\%}
\alias{\%===\%}
\alias{\%!=\%}
\alias{\%$\%}
\alias{\%!$\%}
\title{Grapes of Wrath}
\usage{
x \%!in\% y

x \%||\% y

x \%==\% y

x \%===\% y

x \%!=\% y

x \%$\% y

x \%!$\% y
}
\description{
Commonly referred to as "binary operators", functions that have a left and
right side are typically defined using \code{\%\%} symbols. Thus the colloquial
term "grapes". The functions below represent a growing list of useful
binary operators that come in handy when working in an R environment.
}
\section{Functions}{
\itemize{
\item \code{\%!in\%}: A friendly version of \code{!x \%in\% y}.

\item \code{\%||\%}: A friendly version of \code{\link[=ifelse]{ifelse()}}

\item \code{\%==\%}: test equality of R objects

\item \code{\%===\%}: Test identical of R objects

\item \code{\%!=\%}: easily test inequality of R objects

\item \code{\%$\%}: alias for \code{x[x \%in\% y]}.

\item \code{\%!$\%}: alias for \code{x[!x \%in\% y]}.
}}

