% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapes.R
\name{grapes}
\alias{grapes}
\alias{\%!in\%}
\alias{\%||\%}
\alias{\%null\%}
\alias{\%==\%}
\alias{\%===\%}
\alias{\%!=\%}
\alias{\%$\%}
\alias{\%!$\%}
\title{The Grapevine}
\usage{
x \%!in\% y

x \%||\% y

x \%null\% y

x \%==\% y

x \%===\% y

x \%!=\% y

x \%$\% y

x \%!$\% y
}
\arguments{
\item{x, y}{The left and right hand side of the grape bunch (binary operator).}
}
\description{
Commonly referred to as "binary operators", functions that have a left and
right side are typically defined using \code{\%\%} symbols. Thus the colloquial
term "grapes". The functions below represent a growing list of useful
binary operators that come in handy when working in an R environment.
}
\section{Functions}{
\itemize{
\item \code{\%!in\%}: A friendly version of \code{!x \%in\% y}.

\item \code{\%||\%}: A friendly version of \code{\link[=ifelse]{ifelse()}}

\item \code{\%null\%}: If \code{x = NULL} or \code{x = ""}, return \code{NULL}, otherwise \code{y}.

\item \code{\%==\%}: Test equality of 2 R objects, minus the attributes.

\item \code{\%===\%}: Test absolute equality of 2 R objects.

\item \code{\%!=\%}: Test absolute inequalit of 2 R objects

\item \code{\%$\%}: Shorthand alias for \code{x[x \%in\% y]}.

\item \code{\%!$\%}: Shorthand alias for \code{x[!x \%in\% y]}, i.e. \code{setdiff(x, y)}.
}}

